Prof. Wieland's request: Store Command Window in Output
Prof. Wieland has requested to store the command window in the output folder as he found it difficult to check if all the models ran through, and complained that it has been cleared by the NK_ET14  (Ellison and Tischbirek (2014) Unconventional government   purchases as a supplement to conventional monetary policy) model. This commit adds the diary functionality to the platform, the command window is stored in the ouput folder under the same name as the xls but in simple text form.